improvement: memisahkan register login dan logout ke controller yg berbeda
